refactor(shared): move workflow import placement into shared package
