+++ packages/shared/src/index.ts
@@ -7,6 +7,7 @@ export * from './canvas/duplicateSubgraph'
 export * from './canvas/duplicateToCanvasActions'
 export * from './canvas/imageVersions'
 export * from './canvas/workflowExchange'
+export * from './canvas/workflowImportPlacement'
 export * from './sceneComposer'
 export * from './videoComposition'
 export * from './nodeRefs'
